Extract SuggestionsAdapter
